Make x number of monster creation work
monster.js now returns a global array of monsters. The monsters postion and
motion_direction is created randomly with the randBetween function. I suspect
that the random function is slightly off at least when creating the motion
direction. An avarage of 999 numbers between 0 and 100 returns 48.77122877122877
and an avarage of 1 and 2 returns 1.4764764764764764.

loop.js then apply a forEach on all the objects in the goombaArray that renders
the monsters.